requirements_pythonanywhere.txt
# Core Flask dependencies
Flask==3.1.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
Flask-Migrate==4.0.5

# Database drivers
PyMySQL==1.1.0

# Forms and validation
WTForms==3.0.1
email-validator==2.0.0

# Security
Werkzeug==3.1.1

# Data visualization
plotly==6.3.0
numpy==2.3.2

# Environment management
python-dotenv==1.0.0

# Production server
gunicorn==21.2.0

# Utilities
Pillow==10.0.1
click==8.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
blinker==1.9.0
